+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for gi-gio-hs-list-model
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.1 -- 2021-08-2021
+
+* Remove `stamp` from C struct for the list store.
+* Disable debug logging by default
+
+## 0.1.0.0 -- 2021-08-2021
 
 * First version. Released on an unsuspecting world.